Коммит

+++ app/database/migrations/2015_09_15_033158_CreateVariablesTable.php
@@ -16,7 +16,7 @@ class CreateVariablesTable extends Migration {
 		{
 			$table->increments('id');
 			$table->decimal('residual_value', 10, 2);
-			$table->integer('monthly_rate');
+			$table->decimal('monthly_rate');
 			$table->integer('useful_life');
 			$table->integer('item_id');
 		});
